CEV Loader + test so we know it works

+++ app/src/main/java/msdoilspill/CEV.java
@@ -6,11 +6,11 @@ package msdoilspill;
  */
 public class CEV {
     public static int cellSize_m = 1000;
-    public int windX_ms = 5; // m/s
-    public int windY_ms = 5; // m/s
-    public int currentX_ms = -3; //m/s
-    public int currentY_ms = 6; //m/s
-    public int temperature_K = 300;
+    public double windX_ms = 5; // m/s
+    public double windY_ms = 5; // m/s
+    public double currentX_ms = -3; //m/s
+    public double currentY_ms = 6; //m/s
+    public double temperature_K = 300;
 
     public double shoreline_half_life = 1;
